capture: parser audit fixes + MQTT 5.0 support
Parsers:
- bacnet: handle 254/255 extended-length string markers (ASHRAE 135);
  fix segmented confirmed-request/complex-ACK PDU header parsing
- c1222: safer ASN.1 boundary peek; unregister on malformed/oversized
- camel: cap ASN.1 length-of-length to 4 bytes; parse multi-byte
  invoke lengths
- certs: add recursion-depth cap in certinfo_alt_names to prevent
  stack overflow on deeply nested SAN
- diameter: refactor TCP reassembly to use ArkimeParserBuf_t
- dcerpc, modbus: unregister (not return 0) on unrecoverable state
- erspan, gre: fix misplaced unlikely() around length check
- mqtt: track protocol version from CONNECT so MQTT 5.0
  SUBSCRIBE/UNSUBSCRIBE correctly skip the Properties varint
- synchrophasor: fix uint16_t loop-counter wrap in CFG-3 parsing
- dhcp: drop trailing spaces in two field descriptions
- adb, postgresql, radius, ciscometadata, imap, esp, esis, isis,
  misc: minor cleanups (dead strcmp, misleading UNUSED wrappers)

Core:
- cloud: correctly skip URL scheme in AWS_CONTAINER_CREDENTIALS_FULL_URI
- command, config, scrubspi: guard NULL GError in error-path logs
- drophash: bounds-check keyLen against internal buffer before read
- http: fix sessionId compare length; off-by-one in curl buffer growth
- main: null-check pw/gr in file-permission check
- plugins: free unused path after successful load
- python, reader-pcapoverip, reader-tzsp, tcphealthcheck, wise,
  writer-s3, suricata: small fixes

Tests:
- regenerate tests/pcap/mqtt_synthetic.pcap with spec-compliant
  MQTT 5.0 framing (Properties varints on CONNECT/SUBSCRIBE/PUBLISH)

+++ capture/parsers/synchrophasor.c
@@ -139,7 +139,7 @@ LOCAL void synchrophasor_parse_frame(ArkimeSession_t *session, const uint8_t *da
 
             if (frameType == 5) {
                 // CFG-3: variable length channel names
-                for (uint16_t j = 0; j < phnmr + annmr + dgnmr && !BSB_IS_ERROR(bsb); j++) {
+                for (uint32_t j = 0; j < (uint32_t)phnmr + annmr + dgnmr && !BSB_IS_ERROR(bsb); j++) {
                     uint8_t chNameLen = 0;
                     BSB_IMPORT_u08(bsb, chNameLen);
                     BSB_IMPORT_skip(bsb, chNameLen);
